+++ modules/farm_rothamsted_experiment_research/src/Form/ProposalEntityForm.php
@@ -41,6 +41,7 @@ class ProposalEntityForm extends ResearchEntityForm {
         'title' => $this->t('In-Field Layout'),
         'weight' => 7,
         'fields' => [
+          'requested_location',
           'field_layout',
           'plot_length',
           'plot_width',