+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add status messages in quick forms if required taxonomy hierarchy does not exist. [#360](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/360)
 - Add initial quote field to proposal entity. [#411](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/411)
 - Add agreed quote file field to experiment plan. [#415](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/415)
+- Add filter tabs for experiment plans. [#414](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/414)
 
 ### Changed