+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change non-standard to varying plot sizes. [#431](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/431)
 - Rename Design entity Treatments tab to Statistical Design. [#418](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/418)
 - Move design entity fields from Description to Statistical Design tab. [#417](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/417)
+- Change label and description of design entity name field. [#430](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/430)
 
 ## [2.10.1](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/milestone/16) 2023-04-05